+++ MinimOsd_Extra/0-build.sh
@@ -16,7 +16,8 @@ make_one(){
 
 
 #make_one 'UAVTALK'
-make_one 'MAVLINK' -DDEBUG
+#make_one 'MAVLINK' -DDEBUG
+make_one 'MAVLINK'
 #make_one 'MWII'
 #make_one 'LTM'