+++ libraries/GCS_MAVLink/GCS_MAVLink.cpp
@@ -11,13 +11,10 @@ License as published by the Free Software Foundation; either
 version 2.1 of the License, or (at your option) any later version.
 */
 
-//#include <FastSerial.h>
-//#define HardwareSerial_h
 #include "../SingleSerial/SingleSerial.h"
 #include "compat.h"
 #include "Arduino.h"
 
-//#include "../AP_Common/AP_Common.h"
 #include <GCS_MAVLink.h>