+++ libraries/GCS_MAVLink/generate.sh
@@ -1,4 +1,5 @@
 #!/bin/sh
+exit
 # script to re-generate mavlink C code for APM
 
 mavdir="$(dirname $0)"